Add SummaryView, update SceneController and refactor.
In class SceneController, the suppliers saved in the  attribute "scenes" now create a new FXMLLoader when invoked with get().  Add summary-view.fxml and SummaryView.java.
Refactor (SceneController.Views becomes In class SceneController, the suppliers saved in the  attribute "scenes" now create a new FXMLLoader when invoked with get() and refactor.ViewName ).