Add and TRY to use class Property (draft).
Before than this commit, method firePropertyChange(...) was (ab)used: now a specialized class (Property) is used (it should replace the currently two field used consisting in both the property and its name (saved as static final String) and should provide an universal interface to use and handle PropertyChangeEvent instances), but a lot of todos are added because heavy refactoring is needed. This is just a try to see how the performance of the application are affected.